// src/app/cv/model/header-form.ts
export interface HeaderForm{

  id: string;
  gender: string;
  nom: string;
  prenom: string;
  numero: string;
  rue: string;
  complementaire: string;
  code: string;
  commune: string;
  telephone: string;
  email: string;
  titre: string;
  phrase: string;
 // cvPath: string;
  imagePath: string;

}